Parser/requirements.txt
PyYAML==5.4.1
requests==2.25.1
python-dateutil==2.8.1
pymongo==3.11.2